update to 104

+++ tests/EmptyPathTest.cpp
@@ -51,7 +51,7 @@ static void drawAndTest(
         "%s style[%d] cap[%d] join[%d] antialias[%d]"
         " filltype[%d] ptcount[%d]",
         str, paint.getStyle(), paint.getStrokeCap(), paint.getStrokeJoin(), paint.isAntiAlias(),
-        path.getFillType(), path.countPoints());
+        (int)path.getFillType(), path.countPoints());
     // uncomment this if you want to step in to see the failure
     //        canvas.drawPath(path, p);
   }
